+++ src/CLI.hs
@@ -1,4 +1,6 @@
-module CLI (printUsage, isImagePage, isPoolPage, Count(Limited, Unlimited), getImgCount, hasUnknownSettings, getOutputDirectory,
+module CLI (printUsage, isImagePage, isPoolPage,
+  Count(Limited, Unlimited), countIsLimited, countIsUnlimited,
+  getImgCount, getPageCount, getThreadCount, hasUnknownSettings, getOutputDirectory,
   requestsHelp, getInputFile) where
 
 import Data.List (isPrefixOf)
@@ -7,17 +9,26 @@ import Text.Read (readMaybe)
 import System.Directory (getCurrentDirectory)
 
 data Count = Limited Int | Unlimited
+countIsLimited :: Count -> Bool
+countIsLimited (Limited _) = True
+countIsLimited _ = False
+countIsUnlimited :: Count -> Bool
+countIsUnlimited Unlimited = True
+countIsUnlimited _ = False
 
 printUsage :: IO ()
 printUsage = mapM_ putStr ["\nUsage:\n"
   , "gelbooru-scrape [options] tags/URL\n"
-  , "Tags are a comma-separated list of tags to include in the search\n"
+  , "Tags are a comma-separated list of tags (without spaces) to include in the search\n"
   , "To exclude a tag, prefix it with '-', eg. blue_hair,-long_hair\n"
   , "URL is a single URL for either a single page or a results page\n"
+  , "Note: URLs and tags should be placed in quotes to avoid weird behavior\n"
   , "Options:\n"
   , "    -h = show this help message\n"
+  , "    -t = thread count [defaults to present logical cores]\n"
   , "    -f = path to file containing newline-separated URLs to download\n"
   , "    -c = maximum count of images to get\n"
+  , "    -p = maximum count of pages to get (if both -c and -p are specified, -c takes precedence)\n"
   , "    -o = output directory [default: working dir]\n"]
 
 isPoolPage :: String -> Bool
@@ -34,6 +45,16 @@ isImagePage s = ("https://" `isPrefixOf` s || "http://" `isPrefixOf` s) && isImg
   isImgURL ('&':'s':'=':'v':'i':'e':'w':_) = True
   isImgURL (_:xs) = isImgURL xs
 
+getThreadCount :: [String] -> Count
+getThreadCount args = if findThreadOption args == 0 then Unlimited else Limited $ findThreadOption args
+  where
+  findThreadOption :: [String] -> Int
+  findThreadOption [] = 0
+  findThreadOption ("-t":os) = if null os then
+    errorWithoutStackTrace "Thread count option was without value" else
+      fromMaybe (errorWithoutStackTrace "Cannot convert thread count value to integer") (readMaybe (head os))
+  findThreadOption (_:os) = findThreadOption os
+
 getImgCount :: [String] -> Count
 getImgCount args = if findImgOption args == 0 then Unlimited else Limited $ findImgOption args
   where
@@ -43,6 +64,16 @@ getImgCount args = if findImgOption args == 0 then Unlimited else Limited $ find
     errorWithoutStackTrace "Image count option was without value" else
       fromMaybe (errorWithoutStackTrace "Cannot convert image count value to integer") (readMaybe (head os))
   findImgOption (_:os) = findImgOption os
+
+getPageCount :: [String] -> Count
+getPageCount args = if findPagesOption args == 0 then Unlimited else Limited $ (findPagesOption args) * 42
+  where
+  findPagesOption :: [String] -> Int
+  findPagesOption [] = 0
+  findPagesOption ("-p":os) = if null os then
+    errorWithoutStackTrace "Page image count option was without value" else
+      fromMaybe (errorWithoutStackTrace "Cannot convert page image count value to integer") (readMaybe (head os))
+  findPagesOption (_:os) = findPagesOption os
 
 getOutputDirectory :: [String] -> IO String
 getOutputDirectory args = if null $ findDirOption args then getCurrentDirectory else return $ findDirOption args
@@ -69,6 +100,8 @@ hasUnknownSettings args = go $ take (length args - 1) args
   go [] = (False, "")
   go (o:os)
     | o == "-c" = go os -- Max image count
+    | o == "-p" = go os -- Max page count
+    | o == "-t" = go os -- Thread count
     | o == "-o" = go os -- Output dir
     | o == "-h" = go os -- Show help
     | o == "-f" = go os -- Output file